added two functionality
