+++ src/components/BrowseMovies.jsx
@@ -1,0 +1,455 @@
+import axios from 'axios';
+import React, { useState, useEffect } from 'react';
+
+function BrowseMovies() {
+    const [getMovies, setMovies] = useState([]);
+    const genre = JSON.parse(localStorage.getItem('genre'));
+    const selectedGenre = JSON.parse(localStorage.getItem('selectedGenre'));
+    const filteredGenre = genre.filter(genre => selectedGenre.includes(genre.id));
+    const filteredGenreName = filteredGenre.map(item => item.name);
+
+    const moviesList = [{
+        page: 1,
+        results: [
+            {
+                adult: false,
+                backdrop_path:
+                    "https://image.tmdb.org/t/p/original/FUnAVgaTs5xZWXcVzPJNxd9qGA.jpg",
+                genre_ids: [878, 28, 18, 12],
+                id: 934632,
+                original_language: "en",
+                original_title: "Rebel Moon — Part Two: The Scargiver",
+                overview:
+                    "The rebels gear up for battle against the ruthless forces of the Motherworld as unbreakable bonds are forged, heroes emerge — and legends are made.",
+                popularity: 2018.458,
+                poster_path:
+                    "https://image.tmdb.org/t/p/original/lV70XeG15PNNg3S1mwaoNk3LVHk.jpg",
+                release_date: "2024-04-19",
+                title: "Rebel Moon — Part Two: The Scargiver",
+                video: false,
+                vote_average: 6.16,
+                vote_count: 443,
+            },
+            {
+                adult: false,
+                backdrop_path:
+                    "https://image.tmdb.org/t/p/original/4woSOUD0equAYzvwhWBHIJDCM88.jpg",
+                genre_ids: [28, 27, 53],
+                id: 1096197,
+                original_language: "en",
+                original_title: "No Way Up",
+                overview:
+                    "Characters from different backgrounds are thrown together when the plane they're travelling on crashes into the Pacific Ocean. A nightmare fight for survival ensues with the air supply running out and dangers creeping in from all sides.",
+                popularity: 1979.885,
+                poster_path:
+                    "https://image.tmdb.org/t/p/original/hu40Uxp9WtpL34jv3zyWLb5zEVY.jpg",
+                release_date: "2024-01-18",
+                title: "No Way Up",
+                video: false,
+                vote_average: 6.344,
+                vote_count: 439,
+            },
+            {
+                adult: false,
+                backdrop_path:
+                    "https://image.tmdb.org/t/p/original/qrGtVFxaD8c7et0jUtaYhyTzzPg.jpg",
+                genre_ids: [28, 878, 12],
+                id: 823464,
+                original_language: "en",
+                original_title: "Godzilla x Kong: The New Empire",
+                overview:
+                    "Following their explosive showdown, Godzilla and Kong must reunite against a colossal undiscovered threat hidden within our world, challenging their very existence – and our own.",
+                popularity: 1722.61,
+                poster_path:
+                    "https://image.tmdb.org/t/p/original/gmGK5Gw5CIGMPhOmTO0bNA9Q66c.jpg",
+                release_date: "2024-03-27",
+                title: "Godzilla x Kong: The New Empire",
+                video: false,
+                vote_average: 6.6,
+                vote_count: 752,
+            },
+            {
+                adult: false,
+                backdrop_path:
+                    "https://image.tmdb.org/t/p/original/1XDDXPXGiI8id7MrUxK36ke7gkX.jpg",
+                genre_ids: [16, 28, 10751, 35, 14],
+                id: 1011985,
+                original_language: "en",
+                original_title: "Kung Fu Panda 4",
+                overview:
+                    "Po is gearing up to become the spiritual leader of his Valley of Peace, but also needs someone to take his place as Dragon Warrior. As such, he will train a new kung fu practitioner for the spot and will encounter a villain called the Chameleon who conjures villains from the past.",
+                popularity: 1513.697,
+                poster_path:
+                    "https://image.tmdb.org/t/p/original/kDp1vUBnMpe8ak4rjgl3cLELqjU.jpg",
+                release_date: "2024-03-02",
+                title: "Kung Fu Panda 4",
+                video: false,
+                vote_average: 7.15,
+                vote_count: 1288,
+            },
+            {
+                adult: false,
+                backdrop_path:
+                    "https://image.tmdb.org/t/p/original/qnVm19Vu2Sc14LoEj82pmqAYr3p.jpg",
+                genre_ids: [12, 28, 18],
+                id: 845111,
+                original_language: "fr",
+                original_title: "Les trois mousquetaires : Milady",
+                overview:
+                    "D'Artagnan, on a quest to rescue the abducted Constance, runs into the mysterious Milady de Winter again. The tension between the Catholics and the Protestants finally escalates, as the king declares war — forcing the now four musketeers into battle. But as the war goes on, they are tested physically, mentally and emotionally.",
+                popularity: 1025.73,
+                poster_path:
+                    "https://image.tmdb.org/t/p/original/rtosxP5sXuoRFPH4sVbMccLIPiV.jpg",
+                release_date: "2023-12-13",
+                title: "The Three Musketeers: Milady",
+                video: false,
+                vote_average: 6.492,
+                vote_count: 326,
+            },
+            {
+                adult: false,
+                backdrop_path:
+                    "https://image.tmdb.org/t/p/original/fUC5VsQcU3m6zmYMD96R7RqPuMn.jpg",
+                genre_ids: [28, 80, 53],
+                id: 1105407,
+                original_language: "en",
+                original_title: "Damaged",
+                overview:
+                    "A Chicago detective travels to Scotland after an emerging serial killer’s crimes match those that he investigated five years earlier, one of which was the crime scene of his murdered girlfriend.",
+                popularity: 833.01,
+                poster_path:
+                    "https://image.tmdb.org/t/p/original/tMO0YLXgJZBnIAjoTSz26zE33YN.jpg",
+                release_date: "2024-04-12",
+                title: "Damaged",
+                video: false,
+                vote_average: 4.724,
+                vote_count: 38,
+            },
+            {
+                adult: false,
+                backdrop_path:
+                    "https://image.tmdb.org/t/p/original/keVfqCMKmJ55nHzmqR2Q5K7LwJt.jpg",
+                genre_ids: [28, 53, 10752],
+                id: 784651,
+                original_language: "hi",
+                original_title: "फाइटर",
+                overview:
+                    "Top IAF aviators come together in the face of imminent danger, to form Air Dragons. Fighter unfolds their camaraderie, brotherhood and battles, internal and external.",
+                popularity: 739.031,
+                poster_path:
+                    "https://image.tmdb.org/t/p/original/zDZowwb9GZGEctAu2PCpjiPQAMM.jpg",
+                release_date: "2024-01-24",
+                title: "Fighter",
+                video: false,
+                vote_average: 5.125,
+                vote_count: 44,
+            },
+            {
+                adult: false,
+                backdrop_path:
+                    "https://image.tmdb.org/t/p/original/7ZP8HtgOIDaBs12krXgUIygqEsy.jpg",
+                genre_ids: [878, 28, 14, 12],
+                id: 601796,
+                original_language: "ko",
+                original_title: "외계+인 1부",
+                overview:
+                    "Gurus in the late Goryeo dynasty try to obtain a fabled, holy sword, and humans in 2022 hunt down an alien prisoner that is locked in a human's body. The two parties cross paths when a time-traveling portal opens up.",
+                popularity: 718.79,
+                poster_path:
+                    "https://image.tmdb.org/t/p/original/8QVDXDiOGHRcAD4oM6MXjE0osSj.jpg",
+                release_date: "2022-07-20",
+                title: "Alienoid",
+                video: false,
+                vote_average: 6.813,
+                vote_count: 288,
+            },
+            {
+                adult: false,
+                backdrop_path:
+                    "https://image.tmdb.org/t/p/original/pwGmXVKUgKN13psUjlhC9zBcq1o.jpg",
+                genre_ids: [28, 14],
+                id: 634492,
+                original_language: "en",
+                original_title: "Madame Web",
+                overview:
+                    "Forced to confront revelations about her past, paramedic Cassandra Webb forges a relationship with three young women destined for powerful futures...if they can all survive a deadly present.",
+                popularity: 709.958,
+                poster_path:
+                    "https://image.tmdb.org/t/p/original/rULWuutDcN5NvtiZi4FRPzRYWSh.jpg",
+                release_date: "2024-02-14",
+                title: "Madame Web",
+                video: false,
+                vote_average: 5.616,
+                vote_count: 1137,
+            },
+            {
+                adult: false,
+                backdrop_path:
+                    "https://image.tmdb.org/t/p/original/oe7mWkvYhK4PLRNAVSvonzyUXNy.jpg",
+                genre_ids: [28, 53],
+                id: 359410,
+                original_language: "en",
+                original_title: "Road House",
+                overview:
+                    "Ex-UFC fighter Dalton takes a job as a bouncer at a Florida Keys roadhouse, only to discover that this paradise is not all it seems.",
+                popularity: 664.823,
+                poster_path:
+                    "https://image.tmdb.org/t/p/original/bXi6IQiQDHD00JFio5ZSZOeRSBh.jpg",
+                release_date: "2024-03-08",
+                title: "Road House",
+                video: false,
+                vote_average: 7.03,
+                vote_count: 1562,
+            },
+            {
+                adult: false,
+                backdrop_path:
+                    "https://image.tmdb.org/t/p/original/2KGxQFV9Wp1MshPBf8BuqWUgVAz.jpg",
+                genre_ids: [16, 28, 12, 35, 10751],
+                id: 940551,
+                original_language: "en",
+                original_title: "Migration",
+                overview:
+                    "After a migrating duck family alights on their pond with thrilling tales of far-flung places, the Mallard family embarks on a family road trip, from New England, to New York City, to tropical Jamaica.",
+                popularity: 641.042,
+                poster_path:
+                    "https://image.tmdb.org/t/p/original/ldfCF9RhR40mppkzmftxapaHeTo.jpg",
+                release_date: "2023-12-06",
+                title: "Migration",
+                video: false,
+                vote_average: 7.526,
+                vote_count: 1236,
+            },
+            {
+                adult: false,
+                backdrop_path:
+                    "https://image.tmdb.org/t/p/original/unCdljyU2FuUBDUcdZezeI9yaaj.jpg",
+                genre_ids: [878, 9648, 53, 28],
+                id: 720321,
+                original_language: "en",
+                original_title: "Breathe",
+                overview:
+                    "Air-supply is scarce in the near future, forcing a mother and daughter to fight for survival when two strangers arrive desperate for an oxygenated haven.",
+                popularity: 606.104,
+                poster_path:
+                    "https://image.tmdb.org/t/p/original/gPAW0a24TrkafrJQ1axEv0I3jov.jpg",
+                release_date: "2024-04-04",
+                title: "Breathe",
+                video: false,
+                vote_average: 5.1,
+                vote_count: 55,
+            },
+            {
+                adult: false,
+                backdrop_path:
+                    "https://image.tmdb.org/t/p/original/ilK1JRbMjo4sMJtNKLOKnqRf1RH.jpg",
+                genre_ids: [10752, 28, 36],
+                id: 976906,
+                original_language: "en",
+                original_title: "Spitfire Over Berlin",
+                overview:
+                    "August 1944. With the American Eighth Air Force poised to strike over Nazi Germany, British Intelligence learns that they could be flying into a deadly trap. With only hours to spare, Flight Lieutenant Edward Barnes must fly a life and death mission over Berlin in his unarmed Spitfire to obtain photographic evidence and save the lives of 1200 men.",
+                popularity: 547.208,
+                poster_path:
+                    "https://image.tmdb.org/t/p/original/xtPPOPTad1qopK6uDe3VlYUa22o.jpg",
+                release_date: "2022-05-13",
+                title: "Spitfire Over Berlin",
+                video: false,
+                vote_average: 3.4,
+                vote_count: 11,
+            },
+            {
+                adult: false,
+                backdrop_path:
+                    "https://image.tmdb.org/t/p/original/uv2twFGMk2qBdyJBJAVcrpRtSa9.jpg",
+                genre_ids: [10752, 28, 18],
+                id: 929590,
+                original_language: "en",
+                original_title: "Civil War",
+                overview:
+                    "In the near future, a group of war journalists attempt to survive while reporting the truth as the United States stands on the brink of civil war.",
+                popularity: 546.424,
+                poster_path:
+                    "https://image.tmdb.org/t/p/original/sh7Rg8Er3tFcN9BpKIPOMvALgZd.jpg",
+                release_date: "2024-04-10",
+                title: "Civil War",
+                video: false,
+                vote_average: 7.437,
+                vote_count: 399,
+            },
+            {
+                adult: false,
+                backdrop_path:
+                    "https://image.tmdb.org/t/p/original/dcnSWFCtk4b2aIzkhq6IDbzoIf1.jpg",
+                genre_ids: [28, 35],
+                id: 942047,
+                original_language: "en",
+                original_title: "Outsource",
+                overview:
+                    "A police chief hires an old friend, who is an international spy, to help him search for a wanted suspect in the Philippines. When the chief dies, all evidence points towards the spy, and he must go to extremes to defend himself.",
+                popularity: 519.856,
+                poster_path:
+                    "https://image.tmdb.org/t/p/original/zIAF0UXtCJTJOYNYWiBfyifaaOi.jpg",
+                release_date: "2022-01-18",
+                title: "Outsource",
+                video: false,
+                vote_average: 2.273,
+                vote_count: 11,
+            },
+            {
+                adult: false,
+                backdrop_path:
+                    "https://image.tmdb.org/t/p/original/qekky2LbtT1wtbD5MDgQvjfZQ24.jpg",
+                genre_ids: [28, 53],
+                id: 984324,
+                original_language: "fr",
+                original_title: "Le salaire de la peur",
+                overview:
+                    "When an explosion at an oil well threatens hundreds of lives, a crack team is called upon to make a deadly desert crossing with nitroglycerine in tow.",
+                popularity: 515.465,
+                poster_path:
+                    "https://image.tmdb.org/t/p/original/jFK2ZLQUzo9pea0jfMCHDfvWsx7.jpg",
+                release_date: "2024-03-28",
+                title: "The Wages of Fear",
+                video: false,
+                vote_average: 5.761,
+                vote_count: 190,
+            },
+            {
+                adult: false,
+                backdrop_path:
+                    "https://image.tmdb.org/t/p/original/fypydCipcWDKDTTCoPucBsdGYXW.jpg",
+                genre_ids: [878, 12, 28],
+                id: 653346,
+                original_language: "en",
+                original_title: "Kingdom of the Planet of the Apes",
+                overview:
+                    "Several generations in the future following Caesar's reign, apes are now the dominant species and live harmoniously while humans have been reduced to living in the shadows. As a new tyrannical ape leader builds his empire, one young ape undertakes a harrowing journey that will cause him to question all that he has known about the past and to make choices that will define a future for apes and humans alike.",
+                popularity: 471.896,
+                poster_path:
+                    "https://image.tmdb.org/t/p/original/fAjyUDKABSehK6pHWvh9ket5X3z.jpg",
+                release_date: "2024-05-08",
+                title: "Kingdom of the Planet of the Apes",
+                video: false,
+                vote_average: 0,
+                vote_count: 0,
+            },
+            {
+                adult: false,
+                backdrop_path:
+                    "https://image.tmdb.org/t/p/original/9c0lHTXRqDBxeOToVzRu0GArSne.jpg",
+                genre_ids: [878, 28],
+                id: 935271,
+                original_language: "en",
+                original_title: "After the Pandemic",
+                overview:
+                    "Set in a post-apocalyptic world where a global airborne pandemic has wiped out 90% of the Earth's population and only the young and immune have endured as scavengers. For Ellie and Quinn, the daily challenges to stay alive are compounded when they become hunted by the merciless Stalkers.",
+                popularity: 409.79,
+                poster_path:
+                    "https://image.tmdb.org/t/p/original/p1LbrdJ53dGfEhRopG71akfzOVu.jpg",
+                release_date: "2022-03-01",
+                title: "After the Pandemic",
+                video: false,
+                vote_average: 4.798,
+                vote_count: 57,
+            },
+            {
+                adult: false,
+                backdrop_path:
+                    "https://image.tmdb.org/t/p/original/qwK9soQmmJ7kRdjLZVXblw3g7AQ.jpg",
+                genre_ids: [28, 12, 53, 80],
+                id: 7451,
+                original_language: "en",
+                original_title: "xXx",
+                overview:
+                    'Xander Cage is your standard adrenaline junkie with no fear and a lousy attitude. When the US Government "recruits" him to go on a mission, he\'s not exactly thrilled. His mission: to gather information on an organization that may just be planning the destruction of the world, led by the nihilistic Yorgi.',
+                popularity: 409.57,
+                poster_path:
+                    "https://image.tmdb.org/t/p/original/xeEw3eLeSFmJgXZzmF2Efww0q3s.jpg",
+                release_date: "2002-08-09",
+                title: "xXx",
+                video: false,
+                vote_average: 5.93,
+                vote_count: 4015,
+            },
+            {
+                adult: false,
+                backdrop_path:
+                    "https://image.tmdb.org/t/p/original/atPlFdUrQl2U9MtUwujrrjnQHBA.jpg",
+                genre_ids: [28, 80, 18, 53],
+                id: 654739,
+                original_language: "ko",
+                original_title: "발신제한",
+                overview:
+                    "On his way to work, a bank manager receives an anonymous call claiming there's a bomb under his car seat, and if anyone exits the car, it will explode unless he can pay a ransom.",
+                popularity: 389.651,
+                poster_path:
+                    "https://image.tmdb.org/t/p/original/y2Aimt8isimtigec3e4kB2G9FMR.jpg",
+                release_date: "2021-06-23",
+                title: "Hard Hit",
+                video: false,
+                vote_average: 7.693,
+                vote_count: 171,
+            },
+        ],
+        total_pages: 1980,
+        total_results: 39588,
+    }];
+
+    function fetchAllGenreList() {
+        const options = {
+            method: 'GET',
+            url: 'https://advanced-movie-search.p.rapidapi.com/genre/movie/list',
+            headers: {
+                'X-RapidAPI-Key': '1559d32918msh582cadeb292b64fp1f0e8ajsn4e82e1148d86',
+                'X-RapidAPI-Host': 'advanced-movie-search.p.rapidapi.com'
+            }
+        };
+        axios(options).then(response => {
+            const matchedGenres = response.data.genres.filter(genre => filteredGenreName.includes(genre.name));
+            matchedGenres.map(genre => searchByGenres(genre.id));
+        }).catch(error => {
+            console.error('Error:', error);
+        });
+    }
+
+    function searchByGenres(genreId) {
+        const options = {
+            method: 'GET',
+            url: 'https://advanced-movie-search.p.rapidapi.com/discover/movie',
+            params: { with_genres: genreId, page: '1' },
+            headers: {
+                'X-RapidAPI-Key': '1559d32918msh582cadeb292b64fp1f0e8ajsn4e82e1148d86',
+                'X-RapidAPI-Host': 'advanced-movie-search.p.rapidapi.com'
+            }
+        };
+        axios(options).then(response => {
+            setMovies(response)
+        }).catch(error => {
+            console.error('Error:', error);
+        });
+    }
+
+    useEffect(() => {
+        fetchAllGenreList()
+    }, []);
+
+    console.log(getMovies)
+
+    // return (
+    //     <div className='bg-black pt-20'>
+    //         {filteredGenreName.map((name, key) => {
+    //             return (
+    //                 <div className='border border-red-500 mb-10 w-full container mx-auto' key={key}>
+    //                     <h4 className='text-gray-300'>{name}</h4>
+    //                     {getMovies.map((movies, key) => (
+    //                         <p className='text-white' key={key}>{movies.original_title}</p>
+    //                     ))}
+    //                 </div>
+    //             );
+    //         })}
+    //     </div>
+    // )
+}
+
+export default BrowseMovies
