--- src/components/index.jsx
@@ -1,9 +1,0 @@
-import Form from "./Home/Form";
-import Chip from "./Movies/Chip";
-import Box from "./Movies/Box";
-import UserInfo from "./Dashboard/UserInfo";
-import Notes from "./Dashboard/Notes";
-import News from "./Dashboard/News";
-import Clock from "./Dashboard/Clock";
-import Weather from "./Dashboard/Weather";
-export { Clock, News, Notes, Form, Chip, Box, UserInfo, Weather };
