+++ src/main.jsx
@@ -1,22 +1,9 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-import "./index.css";
-import Home from "./pages/Home.jsx";
-import Movies from "./pages/Movies.jsx";
-import { BrowserRouter, Route, Routes } from "react-router-dom";
-import NotFound from "./pages/404.jsx";
-import Browse from "./pages/Browse.jsx";
-import Dashboard from "./pages/Dashboard.jsx";
-ReactDOM.createRoot(document.getElementById("root")).render(
+import React from 'react'
+import ReactDOM from 'react-dom/client'
+import App from './App.jsx'
+
+ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
-    <BrowserRouter>
-      <Routes>
-        <Route path="/" element={<Home />} />
-        <Route path="/movies" element={<Movies />} />
-        <Route path="/dashboard" element={<Dashboard />} />
-        <Route path="/browse" element={<Browse />} />
-        <Route path="/*" element={<NotFound />} />
-      </Routes>
-    </BrowserRouter>
-  </React.StrictMode>
-);
+    <App />
+  </React.StrictMode>,
+)
