--- src/pages/404.jsx
@@ -1,3 +1,0 @@
-export default function NotFound() {
-  return <p>404 - Page not found</p>;
-}
