--- src/pages/Home.jsx
@@ -1,8 +1,0 @@
-import { Form } from "../components";
-export default function Home() {
-  return (
-    <>
-      <Form />
-    </>
-  );
-}
